Rebuild images only on the weekend or when merges happen.
Previously, images were rebuilt every day, which increases the number of
stored images in the cern registry to two to three depending on popularity.
This prevents us from rebuilding them on the day if desired.

Here, the rebuild is reduced to once per week, letting us start the week
with up-to-date images. If we decide to rebuild on a day, there should
be enough space to store the second wave of images.